Import live.2011.11.29.tar.gz
2011.11.28
- Use "uintptr_t" and "intptr_t" in a few places, where appropriate, to eliminate compiler warnings when compiing for 64-bits.
  (Thanks to the VLC developers for this suggestion.)
- When we create each (datagram or stream) socket, set the 'close on exec' property (if we can).
  (Thanks to the VLC developers for this suggestion.)

2011.11.29
- Updated "MPEG2TransportStreamMultiplexor" to change the Program Map PID from 0x10 to 0x30.  (Thanks to Chris Richardson,
  who noted that PIDs below 0x30 are not valid for ATSC or DVB (although they may be valid for other Transport Streams).)
- Updated the "NetAddressList()" constructor (for resolving names->IP addresses) to use "getaddrinfo()" - if available - rather
  than the older, deprecated "gethostbyname()" function.  Also, the "ourIPAddress()" code (in "GroupsockHelper.cpp") now uses
  the "NetAddressList()" code (if the initial multicast loopback method fails) to look up our host name.

+++ BasicUsageEnvironment/BasicHashTable.cpp
@@ -263,10 +263,10 @@ unsigned BasicHashTable::hashIndexFromKey(char const* key) const {
     }
     result &= fMask;
   } else if (fKeyType == ONE_WORD_HASH_KEYS) {
-    result = randomIndex((unsigned long)key);
+    result = randomIndex((uintptr_t)key);
   } else {
     unsigned* k = (unsigned*)key;
-    unsigned long sum = 0;
+    uintptr_t sum = 0;
     for (int i = 0; i < fKeyType; ++i) {
       sum += k[i];
     }
@@ -275,4 +275,3 @@ unsigned BasicHashTable::hashIndexFromKey(char const* key) const {
 
   return result;
 }
-
